neural stack rewrite
